% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\docType{package}
\name{fars-package}
\alias{fars}
\alias{fars-package}
\title{fars: fars package assignment for 4th week of the coursera JH}
\description{
This package is an assignment of coursera::JH R on Fars topic.
}
\author{
\strong{Maintainer}: Ines Teaca \email{ines.asandei@gmail.com}

}
\keyword{internal}
