% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\docType{package}
\name{fars2-package}
\alias{fars2}
\alias{fars2-package}
\title{fars2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
