% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_summarize_years}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer state.num number}

\item{year}{an integer}
}
\value{
This function returns a list of tibbles
}
\description{
Read in csv for  \code{year} and plot all the accidents
in \code{state} on a map.
}
\examples{
fars_map_state(1, 2013)
}
