% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\name{fars_read}
\alias{fars_read}
\title{filename}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the name of the csv to be loaded}
}
\value{
This function returns a tibble corresponding to the csv
}
\description{
Function to read files for farst
}
\examples{
acc13 <- fars_read("accident_2013.csv.bz2")
}
