% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Is an integer values needed to build the file in accordance with affiliate year}
}
\value{
This function returns a list of tibbles
}
\description{
Read in csv for eah year in \code{years} and returns the
months and year in that csv
}
\examples{
fars_read_years(c(2013,2014,2015))
}
