% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer}
}
\value{
This function returns a tibble that summarises each of the csv
iles corresponding to the \code{years}
}
\description{
summarises the number of entries in the csv files for each
year and month in \code{years}
}
\examples{
fars_summarize_years(c(2013,2014,2015))
}
