% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2-package.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer year (2013-2015)}
}
\value{
This function returns a string that is
the file name of a csv
}
\description{
Function to build file names readable
}
\examples{
make_filename(2013)
}
